+++ LandArmy.h
@@ -7,7 +7,7 @@ public:
 	LandArmy(const Player &owner, Territory *location, int strength);
 	~LandArmy() = default;
 
-	void attack(LandArmy &defendingArmy, float defenceMultiplier);
+	void attack(LandArmy &defendingArmy, double defenceMultiplier);
 	virtual void move(Territory *location, int strength) override;
 
 private:
